// src/core/types/IAdminGetAllUser.ts
export interface IAdminAllUsers {
  createdAt: string;
  created_at: string;
  email: string;
  emailVerified: boolean;
  firstName: string;
  fullName: string;
  id: number;
  lastName: string;
  membershipDate: null;
  phoneNumber: string;
  profilePicture: null;
  resetCode: null;
  resetCodeExpires: null;
  role: string;
  updatedAt: string;
  updated_at: string;
  verificationCode: string;
  verificationCodeExpires: string;
}